+++ src/hyperbase/constants.py
@@ -1,4 +1,7 @@
 # Pre-defined system atoms
 compound_noun_builder = "+/B/."
 possessive_builder = "poss/Bp.am/."
-list_or_matches_builder = "list/J/."
+list_of_matches_builder = "list/J/."
+
+# Pattern functions
+PATTERN_FUNCTIONS: set[str] = {"var", "atoms", "lemma", "any"}